Add files via upload
Made 2 versions of Transform and I can't make the merge file with SQL. If we can get the "netflix merge file though SQL, I think it would be better. Please see it and advise to me.

On the "ETL-Transform_Main" file, I did the merge & some cleaning works. If we can do anything with SQL, you can update this code on Main file.